add .env, export in other files, default values on update, 404 exception on delete

+++ libraryapi/src/main/java/com/library/api/libraryapi/services/AuthorService.java
@@ -36,7 +36,10 @@ public class AuthorService {
         return authorOpt.get();
     }
 
-    public void deleteAuthorById(UUID id) {
+    public void deleteAuthorById(UUID id) throws NotFoundException {
+        if (!authorRepository.existsById(id)) {
+            throw new NotFoundException();
+        }
         authorRepository.deleteById(id);
     }
